backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
pydantic
google-generativeai
python-dotenv
